+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-- Change default key to `V` to not conflict with friend list
+- BREAKING: Rename `toggleAutoJump` in config to `autoJump`
+- Feature: Localize (de)activation message ([#40](https://github.com/emonadeo/autorun/issues/40))
+- Remove niche languages to drastically reduce artifact size
+- Replace `config.autorun.toggleAutoJump` with vanilla `options.autoJump`
+- Fix tooltip of `alwaysSprint` showing the title instead of description
 
-Full Changelog: [`mc26.1-v1.5.0...mc26.2-v1.6.0`](https://github.com/emonadeo/autorun/compare/mc26.1-v1.5.0...mc26.2-v1.6.0)
+Full Changelog: [`mc26.2-v1.6.0...mc26.2-v2.0.0`](https://github.com/emonadeo/autorun/compare/mc26.2-v1.6.0...mc26.2-v2.0.0)